refactor: Improve network layer structure and visibility
This commit refactors the network layer by:

- Introducing `SupabaseEndpoint.kt` to define API endpoints, currently including a `TestTable` object.
- Moving `Klient.kt` from the `helpers` package to the `network` package and making it `internal`.
- Making `KtorClientProvider` `internal`.
- Making the `logger()` function in `Logger.kt` `internal`.
- Adding a placeholder `BirdieApi.kt` object.
