+++ birdie-lib/src/commonMain/kotlin/ke/don/birdie_lib/helpers/Logger.kt
@@ -2,6 +2,6 @@ package ke.don.birdie_lib.helpers
 
 import co.touchlab.kermit.Logger
 
-inline fun <reified T> logger(): Logger {
+internal inline fun <reified T> logger(): Logger {
     return Logger.withTag(T::class.simpleName ?: "Birdie")
 }
