+++ birdie-lib/src/commonMain/kotlin/ke/don/birdie_lib/network/api/BirdieApi.kt
@@ -1,0 +1,5 @@
+package ke.don.birdie_lib.network.api
+
+object BirdieApi {
+
+}
